% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_summary_demographics.R
\name{cohorts_summary_demographics}
\alias{cohorts_summary_demographics}
\title{Creates demographic summary table}
\usage{
cohorts_summary_demographics(new.data.set)
}
\arguments{
\item{new.data.set}{the dataset to be summarized}
}
\value{
a table of demographics for a given dataset
}
\description{
Creates demographic summary table
}
