% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condition_query_sql}
\alias{condition_query_sql}
\title{This query represents dataset "t1d_cohort_example_disease_concept_set" for domain "condition" and was generated for All of Us Controlled Tier Dataset v7
January 16, 2024}
\format{
\subsection{\code{condition_query_sql} is a character vector with 1 element}{
}
}
\usage{
condition_query_sql
}
\description{
This query represents dataset "t1d_cohort_example_disease_concept_set" for domain "condition" and was generated for All of Us Controlled Tier Dataset v7
January 16, 2024
}
\keyword{datasets}
