% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_dataset_with_OMOP_CONCEPT_ID.R
\name{create_condition_dataset_with_OMOP_CONCEPT_ID}
\alias{create_condition_dataset_with_OMOP_CONCEPT_ID}
\title{Create condition dataset with OMOP_CONCEPT_ID}
\usage{
create_condition_dataset_with_OMOP_CONCEPT_ID(
  OMOPCONCEPTID,
  query_sql = condition_query_sql
)
}
\arguments{
\item{OMOPCONCEPTID}{The OMOP_CONCEPT_ID of the condition to be queried.}

\item{query_sql}{The SQL query to be used to create the dataset.}
}
\value{
A list containing the OMOPCONCEPTID, query_sql, and query_result_path.
}
\description{
This function creates a dataset of condition data with a specific OMOP_CONCEPT_ID.
}
