% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_drug_exposure_dataset_with_OMOP_CONCEPT_ID.R
\name{create_drug_exposure_dataset_with_OMOP_CONCEPT_ID}
\alias{create_drug_exposure_dataset_with_OMOP_CONCEPT_ID}
\title{create a dataset with drug exposure data for a given OMOP concept ID}
\usage{
create_drug_exposure_dataset_with_OMOP_CONCEPT_ID(
  OMOPCONCEPTID,
  query_sql = drug_exposure_query_sql
)
}
\arguments{
\item{OMOPCONCEPTID}{the OMOP concept ID for the drug exposure}

\item{query_sql}{the SQL query to use to create the dataset}
}
\value{
a list with the OMOP concept ID, the query SQL, and the path to the query results
}
\description{
create a dataset with drug exposure data for a given OMOP concept ID
}
\examples{
OMOPCONCEPTID <- 21600713
tmp <- create_drug_exposure_dataset_with_OMOP_CONCEPT_ID(OMOPCONCEPTID)
tmp.person_ids <- get_person_ids_with_drug_exposure(tmp$query_result_path)
}
