% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_measurement_dataset_with_multiple_OMOP_CONCEPT_IDs.R
\name{create_measurement_dataset_with_multiple_OMOP_CONCEPT_IDs}
\alias{create_measurement_dataset_with_multiple_OMOP_CONCEPT_IDs}
\title{Create measurement dataset with multiple OMOP_CONCEPT_IDs}
\usage{
create_measurement_dataset_with_multiple_OMOP_CONCEPT_IDs(
  OMOPCONCEPTIDs,
  file_label,
  query_sql = measurement_query_sql
)
}
\arguments{
\item{OMOPCONCEPTIDs}{The OMOP_CONCEPT_IDs of the measurements to be queried.}

\item{file_label}{The label for the file to be created.}

\item{query_sql}{The SQL query to be used to create the dataset. Default is measurement_query_sql.}
}
\value{
A list containing the OMOPCONCEPTID, query_sql, query_result_path, and dataset_measurement_df.
}
\description{
This function creates a dataset of measurement data with multiple OMOP_CONCEPT_IDs.
}
