% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_measurement_dataset_with_OMOP_CONCEPT_ID.R
\name{create_measurement_dataset_with_OMOP_CONCEPT_ID}
\alias{create_measurement_dataset_with_OMOP_CONCEPT_ID}
\title{Create a measurement dataset with OMOP_CONCEPT_ID}
\usage{
create_measurement_dataset_with_OMOP_CONCEPT_ID(
  OMOPCONCEPTID,
  query_sql = measurement_query_sql
)
}
\arguments{
\item{OMOPCONCEPTID}{The OMOP Concept ID of the measurement of interest.}

\item{query_sql}{= measurement_query_sql}
}
\value{
A list containing the OMOPCONCEPTID, query_sql, and query_result_path.
}
\description{
This function takes an OMOP Concept ID in the Measurement domain and returns a dataset with the following columns: person_id, measurement_concept_id, measurement_date, measurement_datetime, measurement_type_concept_id, operator_concept_id, value_as_number, value_as_concept_id, unit_concept_id, range_low, range_high, and visit_occurrence_concept_id.
}
