% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset_procedure_sql}
\alias{dataset_procedure_sql}
\title{procedure_query_sql
This query represents dataset "kidney_procedure_OMOP_4322471" for domain "procedure" and was generated for All of Us Controlled Tier Dataset v7}
\format{
\subsection{\code{procedure_query_sql} is a character vector with 1 element}{
}
}
\usage{
dataset_procedure_sql
}
\description{
procedure_query_sql
This query represents dataset "kidney_procedure_OMOP_4322471" for domain "procedure" and was generated for All of Us Controlled Tier Dataset v7
}
\keyword{datasets}
