% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_to_csv_to_bucket.R
\name{dataset_to_csv_to_bucket}
\alias{dataset_to_csv_to_bucket}
\title{dataset_to_csv_to_bucket}
\usage{
dataset_to_csv_to_bucket(name_to_save)
}
\arguments{
\item{name_to_save}{the name of the dataset to be saved}
}
\value{
a csv file in the bucket
}
\description{
This function takes a dataset and saves it as a csv file in the bucket
}
