% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_exposure_query_sql}
\alias{drug_exposure_query_sql}
\title{SQL query from the All of Us Research Workbench to create the drug exposure dataset}
\format{
\subsection{\code{drug_exposure_query_sql} is a character vector with 1 element}{
}
}
\source{
\url{https://workbench.researchallofus.org/workspaces/aou-rw-c8211e3d/commonmetabolicdiseasegeneticassociationanalysisv7/notebooks/preview/Phenotype\%20Harmonization\%20in\%20the\%20All\%20of\%20Us\%20Cohort\%20v7\%2012Dec2023\%20AKM.ipynb}
}
\usage{
drug_exposure_query_sql
}
\description{
This query was taken 'verbatim' from the All of Us Research Workbench on December 12, 2023.
quote: This query represents dataset "all_insulin_example" for domain "drug" and was generated for All of Us Controlled Tier Dataset v7
Documentation exists in the internally available notebook: Phenotype Harmonization in the All of Us Cohort v7 12Dec2023 AKM.ipynb
}
\keyword{datasets}
