% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_measurement_df.R
\name{get_filtered_measurement_df}
\alias{get_filtered_measurement_df}
\title{get_filtered_measurement_df}
\usage{
get_filtered_measurement_df(
  dataset_measurement_df,
  filter_unit_concept_name = NULL,
  filter_operator_concept_name = NULL,
  filter_visit_occurrence_concept_name = NULL,
  filter_measurement_type_concept_name = NULL
)
}
\arguments{
\item{dataset_measurement_df}{Data frame as output from create_measurement_dataset_with_OMOP_CONCEPT_ID function}

\item{filter_unit_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_operator_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_visit_occurrence_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_measurement_type_concept_name}{list of string values for filtering. example c("value1", "value2")}
}
\value{
A data frame with the filtered values
}
\description{
This function takes an data frame created by the create_measurement_dataset_with_OMOP_CONCEPT_ID function and applies the filters operator_concept_name, unit_concept_name, visit_occurrence_concept_name and measurement_type_concept_name
}
