% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurement_table_summary.R
\name{get_measurement_table_summary}
\alias{get_measurement_table_summary}
\title{get_measaurement_table_summary}
\usage{
get_measurement_table_summary(query_result_path)
}
\arguments{
\item{query_result_path}{The path to the query result.}
}
\value{
A data frame with the following columns: standard_concept_name, measurement_type_concept_name, operator_concept_name, unit_concept_name, visit_occurrence_concept_name, and summary statistics.
}
\description{
This function takes an OMOP Concept ID in the Measurement domain returns a summary of the non-missing values stratified by standard_concept_name, measurement_type_concept_name, operator_concept_name,unit_concept_name, and visit_occurrence_concept_name.
}
