% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_dataset_with_OMOP_CONCEPT_ID.R
\name{get_person_ids_with_condition}
\alias{get_person_ids_with_condition}
\title{Get the IDs of the people in a condition dataset}
\usage{
get_person_ids_with_condition(query_result_path)
}
\arguments{
\item{query_result_path}{The path to the query result.}
}
\value{
A vector of person_ids.
}
\description{
This function returns the person_ids of the people in a condition dataset.
}
