% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_person_ids_with_drug_exposure.R
\name{get_person_ids_with_drug_exposure}
\alias{get_person_ids_with_drug_exposure}
\title{Get person IDs with drug exposure}
\usage{
get_person_ids_with_drug_exposure(query_result_path, exclude_inpatient = FALSE)
}
\arguments{
\item{query_result_path}{The path to the query result}

\item{exclude_inpatient}{Whether to exclude inpatient visits}
}
\value{
A vector of person IDs with drug exposure
}
\description{
This function returns a vector of person IDs with drug exposure
}
