% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_bucket.R
\name{html_to_bucket}
\alias{html_to_bucket}
\title{html_to_bucket}
\usage{
html_to_bucket(name_to_save)
}
\arguments{
\item{name_to_save}{the name of the html file to be saved}
}
\value{
a html file in the bucket
}
\description{
This function takes a html file and saves it as a html file in the bucket
}
