% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manninglabr-package.R
\docType{package}
\name{manninglabr-package}
\alias{manninglabr}
\alias{manninglabr-package}
\title{manninglabr: Manning Lab Tools}
\description{
Tools and utilities built for the Manning lab research team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/akmanning/manninglabr}
  \item Report bugs at \url{https://github.com/akmanning/manninglabr/issues}
}

}
\author{
\strong{Maintainer}: Alisa Manning \email{amanning@broadinstitute.org} (\href{https://orcid.org/0000-0003-0247-902X}{ORCID})

}
\keyword{internal}
