% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurement_query_sql}
\alias{measurement_query_sql}
\title{measurement_query_sql
This query represents dataset "creatinine_mass-volume_in_serum_plasma_LOINC_2160-0_OMOP_3016723" for domain "measurement" and was generated for All of Us Controlled Tier Dataset v7}
\format{
\subsection{\code{measurement_query_sql} is a character vector with 1 element}{
}
}
\usage{
measurement_query_sql
}
\description{
measurement_query_sql
This query represents dataset "creatinine_mass-volume_in_serum_plasma_LOINC_2160-0_OMOP_3016723" for domain "measurement" and was generated for All of Us Controlled Tier Dataset v7
}
\keyword{datasets}
