% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_dataset_with_OMOP_CONCEPT_ID.R
\name{read_condition_bq_export_from_workspace_bucket}
\alias{read_condition_bq_export_from_workspace_bucket}
\title{Read condition dataset from workspace bucket}
\usage{
read_condition_bq_export_from_workspace_bucket(export_path)
}
\arguments{
\item{export_path}{The path to the query result.}
}
\value{
A tibble of condition data.
}
\description{
This function reads a condition dataset from the workspace bucket.
}
