% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/read_drug_exposure_bq_export_from_workspace_bucket.R
\name{read_drug_exposure_bq_export_from_workspace_bucket}
\alias{read_drug_exposure_bq_export_from_workspace_bucket}
\title{Read the data directly from Cloud Storage into memory.}
\usage{
read_drug_exposure_bq_export_from_workspace_bucket(export_path)
}
\arguments{
\item{export_path}{The path to the query result.}
}
\value{
A tibble of drug exposure data.
}
\description{
Read the data directly from Cloud Storage into memory.
}
