% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/read_measurement_bq_export_from_workspace_bucket.R
\name{read_measurement_bq_export_from_workspace_bucket}
\alias{read_measurement_bq_export_from_workspace_bucket}
\title{Read drug exposure dataset from workspace bucket}
\usage{
read_measurement_bq_export_from_workspace_bucket(export_path)
}
\arguments{
\item{export_path}{The path to the query result.}
}
\value{
A tibble of drug exposure data.
}
\description{
This function reads a drug exposure dataset from the workspace bucket.
}
