% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/read_person_df_bq_export_from_workspace_bucket.R
\name{read_person_df_bq_export_from_workspace_bucket}
\alias{read_person_df_bq_export_from_workspace_bucket}
\title{Direct translation of the R function read_bq_export_from_workspace_bucket from the All of Us Research Workbench.`}
\usage{
read_person_df_bq_export_from_workspace_bucket(export_path)
}
\arguments{
\item{export_path}{The path to the export directory in the workspace bucket.}
}
\value{
A data frame with the contents of the export.
This query represents dataset "All Participants" for domain "person" and was generated for All of Us Controlled Tier Dataset v7
}
\description{
Direct translation of the R function read_bq_export_from_workspace_bucket from the All of Us Research Workbench.`
}
