% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/read_procedure_bq_export_from_workspace_bucket.R
\name{read_procedure_bq_export_from_workspace_bucket}
\alias{read_procedure_bq_export_from_workspace_bucket}
\title{Read procedure domain BQ export from workspace bucket}
\usage{
read_procedure_bq_export_from_workspace_bucket(export_path)
}
\arguments{
\item{export_path}{The path to the query result.}
}
\value{
A tibble of procedure data.
}
\description{
Read procedure domain BQ export from workspace bucket
}
