% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_summary_demographics.R
\name{cohorts_summary_demographics}
\alias{cohorts_summary_demographics}
\title{Creates demographic summary table}
\usage{
cohorts_summary_demographics(person_df, true_false_variable = NA)
}
\arguments{
\item{person_df}{the dataset to be summarized}

\item{true_false_variable}{the name of a true/false variable in the dataset to stratify upon}
}
\value{
a table of demographics for a given dataset
}
\description{
Creates demographic summary table
}
