% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurement_table_summary.R
\name{get_measurement_table_summary}
\alias{get_measurement_table_summary}
\title{get_measaurement_table_summary}
\usage{
get_measurement_table_summary(
  dataset_measurement_df,
  filter_missing_values_as_number = FALSE,
  filter_unit_concept_name = NULL,
  filter_operator_concept_name = NULL,
  filter_visit_occurrence_concept_name = NULL,
  filter_measurement_type_concept_name = NULL
)
}
\arguments{
\item{dataset_measurement_df}{Data frame as output from create_measurement_dataset_with_OMOP_CONCEPT_ID function}

\item{filter_missing_values_as_number}{logical value for filtering missing values. Default is FALSE.}

\item{filter_unit_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_operator_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_visit_occurrence_concept_name}{list of string values for filtering. example c("value1", "value2")}

\item{filter_measurement_type_concept_name}{list of string values for filtering. example c("value1", "value2")}
}
\value{
A data frame with the following columns: standard_concept_name, measurement_type_concept_name, operator_concept_name, unit_concept_name, visit_occurrence_concept_name, and summary statistics.
}
\description{
This function takes an OMOP Concept ID in the Measurement domain returns a summary of the non-missing values stratified by standard_concept_name, measurement_type_concept_name, operator_concept_name,unit_concept_name, and visit_occurrence_concept_name.
}
